Reduce LED Matrix to use uint32_t

+++ quantum/eeconfig.h
@@ -44,7 +44,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define EECONFIG_HAPTIC (uint32_t *)24
 
 // Mutually exclusive
-#define EECONFIG_LED_MATRIX (uint64_t *)28
+#define EECONFIG_LED_MATRIX (uint32_t *)28
 #define EECONFIG_RGB_MATRIX (uint64_t *)28
 
 // Size of EEPROM being used for core data storage
